upload round2 step3
