+++ CodeStarterCamp_Week2/main.swift
@@ -2,30 +2,31 @@
 //  main.swift
 //  CodeStarterCamp_Week2
 //
-//  Created by yagom.
-//  Copyright © yagom academy. All rights reserved.
+//  Created by 지현 on 12/14/23.
 //
 
 import Foundation
 
-func makeLottoNumber() -> Set<Int> {
+var lottoRound: Int = 0
+var lottoInfo: Dictionary<String, Set<Int>> = Dictionary<String, Set<Int>>()
+
+func makeNumber() {
     var lottoNumber: Set<Int> = Set<Int>()
+    lottoRound += 1
     while lottoNumber.count < 6 {
         lottoNumber.insert(Int.random(in: 1...45))
     }
-    return lottoNumber
+    lottoInfo["\(lottoRound)회차"] = lottoNumber
 }
 
-
-func compareNumber(lottoNumber: Set<Int>, mylottoNumbers: Array<Int>) {
-    var containNumber: Set<Int> = Set<Int>()
-    containNumber = lottoNumber.intersection(mylottoNumbers)
-    if !(containNumber.isEmpty) {
-        print("축하합니다! 겹치는 번호는 \(containNumber.map{String($0)}.joined(separator:", ")) 입니다!")
-    } else {
-        print("아쉽지만 겹치는 번호가 없습니다.")
+func checkLottoNum(round: Int) {
+    if let lottoNum = lottoInfo["\(lottoRound)회차"] {
+        print("\(round)회차의 로또 당첨 번호는 \(lottoNum.map{String($0)}.joined(separator:", ")) 입니다.")
     }
 }
 
-let myLottoNumbers: [Int] = [1, 2, 3, 4, 5, 6]
-compareNumber(lottoNumber: makeLottoNumber(), mylottoNumbers: myLottoNumbers)
+for _ in 1...5 {
+    makeNumber()
+}
+
+checkLottoNum(round: 2)
